requirements.txt
asgiref==3.4.1
Django==3.2.23
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.1.1
